removing old classes, build errors, simpler dev command
